chore: organize storybook folder
