+++ webui/react/src/hooks/useModal/Checkpoint/useModalCheckpoint.test.tsx
@@ -5,7 +5,7 @@ import React, { useCallback } from 'react';
 import Button from 'components/kit/Button';
 import { StoreProvider as UIProvider } from 'shared/contexts/stores/UI';
 import { ModalCloseReason } from 'shared/hooks/useModal/useModal';
-import { generateTestExperimentData } from 'storybook/shared/generateTestData';
+import { generateTestExperimentData } from 'test_utils/generateTestData';
 
 import useModalCheckpoint, { Props } from './useModalCheckpoint';
