+++ webui/react/src/hooks/useModal/Experiment/useModalExperimentCreate.test.tsx
@@ -4,7 +4,7 @@ import React, { useEffect } from 'react';
 
 import Button from 'components/kit/Button';
 import { setAuth } from 'stores/auth';
-import { generateTestExperimentData } from 'storybook/shared/generateTestData';
+import { generateTestExperimentData } from 'test_utils/generateTestData';
 
 import useModalExperimentCreate, { CreateExperimentType } from './useModalExperimentCreate';
