+++ webui/react/src/hooks/useModal/HyperparameterSearch/useModalHyperparameterSearch.test.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import Button from 'components/kit/Button';
 import { createExperiment as mockCreateExperiment } from 'services/api';
 import { ClusterProvider } from 'stores/cluster';
-import { generateTestExperimentData } from 'storybook/shared/generateTestData';
+import { generateTestExperimentData } from 'test_utils/generateTestData';
 
 import useModalHyperparameterSearch from './useModalHyperparameterSearch';
